requirements.txt
Pillow
instagrapi
bidi.algorithm